% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_num.R
\name{as_num}
\alias{as_num}
\title{Gently coerce to numeric: convert all chars to NA}
\usage{
as_num(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector
}
\description{
Gently coerce to numeric: convert all chars to NA
}
\examples{
x <- c("a", 1, 2); as_num(x)

# Alternate way: 
x <- c("a", 1, 2); class(x) <- "numeric"

}
