% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genvar.R
\name{genvar}
\alias{genvar}
\title{Generalized Variance}
\usage{
genvar(x)
}
\arguments{
\item{x}{data frame}
}
\value{
numeric
}
\description{
Calculates generalized variance
}
\examples{
x <- data.frame(cbind(seq(7,100,10), seq(11,20,1), seq(5,35,7))); genvar(x)
}

