% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goji.R
\docType{package}
\name{goji-package}
\alias{goji-package}
\alias{goji}
\title{goji: Miscellaneous utility functions by G!}
\description{
Recommended for co-authors of G and those trying to use his scripts.
}
\author{
Gaurav Sood
}
