% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grpfn.R
\name{grpfun}
\alias{grpfun}
\title{Group Sum Mean Var}
\usage{
grpfun(var, group, fun)
}
\arguments{
\item{var}{variable}

\item{group}{grouping var}

\item{fun}{can be "var", "mean", "sum"}
}
\value{
vector
}
\description{
Calculate sum, mean, var by group with same n as original dataset
Takes a variable, a group variable, function
}
\examples{
x <- c(1:10); grp <- c(rep(1:5, each=2));grpfun(x, grp, "mean")
}

