% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice.R
\name{nice}
\alias{nice}
\title{nice}
\usage{
nice(x)
}
\arguments{
\item{x}{number of vector of numbers}
}
\value{
Vector of strings
}
\description{
Formats numbers to the preferred format: X.XX
Takes a number or a vector
}
\examples{
x <- c(.200, .123, 2.2); nice(x)
}

