% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nolead0s.R
\name{nolead0s}
\alias{nolead0s}
\title{No Lead 0s}
\usage{
nolead0s(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector of strings
}
\description{
Takes a vector of numerics and strips 0s before the decimal. For axis labeling etc.
}
\examples{
x <- c(0.00,0.1); nolead0s(x)
}
