% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_prop.R
\name{se_prop}
\alias{se_prop}
\title{SE for the proportion}
\usage{
se_prop(p, n)
}
\arguments{
\item{p}{proportion}

\item{n}{size of the vector}
}
\value{
vector
}
\description{
Takes p and n
}
\examples{
se_prop(.50, 100)

}
