% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize}
\usage{
standardize(x)
}
\arguments{
\item{x}{vector of numbers}
}
\value{
vector
}
\description{
Standardize a variable
Takes a vector
}
\examples{
x <- c(1,2,3); standardize(x)
}

