% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{stars}
\alias{stars}
\title{Star Generator}
\usage{
stars(pvalue)
}
\arguments{
\item{pvalue}{vector of p-values}
}
\value{
numeric
}
\description{
Takes p-value or vector of p-values and generates stars
}
\examples{
x <- c(.002, .01, .2); stars(x)
}
