% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Calculate Entropy}
\usage{
entropy(x, n)
}
\arguments{
\item{x}{vector}

\item{n}{}
}
\value{
recoded vector
}
\description{
Takes a vector and number
}
\examples{
x <- c(1,2,3); entropy(x,2)
}

